Added separate methods to map ResultSet to Book in BookDaoImpl. Minor fixes applied.

+++ src/main/resources/init_db.sql
@@ -5,7 +5,7 @@ USE `intro`;
 CREATE TABLE `books` (
     `id` INT NOT NULL AUTO_INCREMENT,
     `title` VARCHAR(255),
-    `price` INT,
+    `price` DECIMAL,
     PRIMARY KEY (`id`)
 );
